requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.23.2
supabase==2.15.1
jinja2==3.1.3
starlette==0.45.3
python-multipart==0.0.20
PyJWT>=2.8.0
bcrypt==4.1.2

qrcode==8.2
reportlab==4.4.2
pillow==11.2.1
cloudinary==1.44.1